requirements.txt
Django==1.6.7
South
Pillow